Optimisation (#1)
* Update main.bf

Swapped V1 and V2 to decrease the amount of movement when carrying digits in the addition. Removed cell 1 (was always set to 1), and shuffled everything else left by 1 to compensate for the new structure.

* Update main.bf

End each loop at the end of the array, to avoid returning to the start and then immediately moving to the end.

* Update main.bf

Printing each cell, which requires a total increment of 48, is now 6*8 (was 5*10-2).

* Update compressed.bf

* Update README.md

* Update main.bf

Removed outdated comment on line 65.